krb-417: Endringer på NestableHierarcy
