+++ src/Workbench/Admin/Need/NeedPage.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { Parentable } from '../../../models/Parentable';
-import NestableHierarcy from '../../../NestableHierarchy/NestableHierarcy';
+import NestableHierarcyWithAccordion from '../../../NestableHierarchy/NestableHierarcyWithAccordion';
 import { INeed } from '../../../Nexus/entities/INeed';
 import { useAppDispatch, useAppSelector } from '../../../store/hooks';
 import {
@@ -27,7 +27,7 @@ function NeedPage(): React.ReactElement {
       <h3 className="mt-3">{t('Needs')}</h3>
 
       <NewNeedForm />
-      <NestableHierarcy
+      <NestableHierarcyWithAccordion
         dispatchfunc={(items: Parentable<INeed>[]) => newNeedList(items)}
         inputlist={project.needs}
         component={<EditNeedForm element={project.needs[0]} />}
