+++ src/Workbench/Components/NestableHiarchySideBar.tsx
@@ -11,6 +11,7 @@ import DragIndicatorIcon from '@mui/icons-material/DragIndicator';
 import { makeStyles, Box } from '@material-ui/core';
 import { Typography } from '@mui/material';
 import theme from '../../theme';
+import NestableHierarcy from '../../NestableHierarchy/NestableHierarcy';
 
 interface IProps<T extends IBaseModel> {
   listOrderFunc: (itemlist: Parentable<T>[]) => void;
@@ -86,55 +87,10 @@ const NestableHierarcySideBar = <T extends IBaseModel>({
   const classes = useStyles();
   const [selected, setSelected] = useState<Parentable<T> | null>(null);
 
-  const convertTreeToList = (tree: Item, key: string, collection: Item[]) => {
-    if ((!tree[key] || tree[key].length === 0) && collection.includes(tree)) {
-      const copiedTree = { ...tree };
-      delete copiedTree.children;
-      collection.push(copiedTree);
-      return;
-    }
-    if (!collection.includes(tree)) {
-      const copiedTree = { ...tree };
-      delete copiedTree.children;
-      collection.push(copiedTree);
-    }
-    for (let i = 0; i < tree[key].length; i += 1) {
-      const child = tree[key][i];
-      child.parent = tree.id === undefined ? '' : tree.id;
-      convertTreeToList(child, key, collection);
-    }
-  };
-
-  const flatten = (list: {
-    items: Item[];
-    dragItem?: Item;
-    targetPath?: number[];
-  }) => {
-    const flattenedCollection: Item[] = [];
-    list.items.forEach((element) => {
-      const copy = element;
-      copy.parent = '';
-      convertTreeToList(element, 'children', flattenedCollection);
-    });
-    return flattenedCollection;
-  };
-
-  const hierarchyList = Utils.parentable2Nestable(inputlist);
-
-  const onChange = (items: {
-    items: Item[];
-    dragItem: Item;
-    targetPath: number[];
-  }) => {
-    const itemList = flatten(items);
-    const returnList: Parentable<T>[] = [];
-    itemList.forEach((elem) => {
-      const clone = { ...elem };
-      delete clone.level;
-      returnList.push(clone as Parentable<T>);
-    });
-    listOrderFunc(returnList);
-  };
+  const { hierarchyList, onChange } = NestableHierarcy(
+    listOrderFunc,
+    inputlist
+  );
 
   const itemClicked = (item: Item) => {
     const selectedParentable = item as Parentable<T>;
